Move activities to /ui/activities package
