+++ Habitica/res/layout/activity_about.xml
@@ -3,7 +3,7 @@
     android:id="@+id/main_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AboutActivity">
+    tools:context=".ui.activities.AboutActivity">
 
     <android.support.design.widget.TabLayout
         android:id="@+id/tab_layout"
