+++ Habitica/res/menu/menu_task_form.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.habitrpg.android.habitica.TaskFormActivity">
+    tools:context=".ui.activities.TaskFormActivity">
     <item android:id="@+id/action_save_changes" android:title="@string/save_changes"
         android:orderInCategory="100" app:showAsAction="always" />
 </menu>
