+++ Habitica/src/com/habitrpg/android/habitica/ui/activities/SkillTasksActivity.java
@@ -1,4 +1,4 @@
-package com.habitrpg.android.habitica;
+package com.habitrpg.android.habitica.ui.activities;
 
 import android.app.Activity;
 import android.content.Intent;
@@ -9,6 +9,7 @@ import android.support.v4.app.FragmentPagerAdapter;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AppCompatActivity;
 
+import com.habitrpg.android.habitica.R;
 import com.habitrpg.android.habitica.ui.adapter.SkillTasksRecyclerViewAdapter;
 import com.habitrpg.android.habitica.ui.fragments.SkillTasksRecyclerViewFragment;
 import com.magicmicky.habitrpgwrapper.lib.models.HabitRPGUser;
